Removed "OPTOPT" make param.
I'm not sure exactly, but I think this targets for a specific architecture. Which means it won't work on anything else? And would really only be needed if you were building on a different machine than you intend to run it on, which is not going to be the case for us and our users.