Remove timidity, freepats depends.
I never could get the MIDI bgm to play, and the OGG tracks work without these two.

Please update if you can get MIDI to work. Apparently it is ONLY included in the shareware version, the registered version doesn't even have it in the GRP file as it just assumes you will be playing with the CD in.