Use $md_id variable (again)
Now that I've changed the module name back to "sorr", some of these refs can go back to `$md_id`.

I don't think it matters much here one way or the other, but the reason I did it no longer exists, so I went ahead and changed it back to how it was before.

For an example when it does matter, see `eduke32.sh` ...then see `ionfury`:

https://github.com/RetroPie/RetroPie-Setup/blob/master/scriptmodules/ports/eduke32.sh

https://github.com/RetroPie/RetroPie-Setup/blob/master/scriptmodules/ports/ionfury.sh

Almost all of the functions in Fury just call the Eduke module's functions. The functions in Eduke, then, *do* need to see the `$md_id` var to know whether they are actually installing Eduke, or if they're just being borrowed to install Fury.